graphsynccontenthandler: use orchestrators
