+++ DFC.ServiceTaxonomy.GraphSync/Handlers/Interfaces/IDeleteOrchestrator.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+using OrchardCore.ContentManagement;
+
+namespace DFC.ServiceTaxonomy.GraphSync.Handlers.Interfaces
+{
+    public interface IDeleteOrchestrator
+    {
+        Task<bool> Unpublish(ContentItem contentItem);
+        Task<bool> Delete(ContentItem contentItem);
+    }
+}
