+++ DFC.ServiceTaxonomy.GraphSync/Handlers/Interfaces/ISyncOrchestrator.cs
@@ -1,0 +1,12 @@
+﻿using System.Threading.Tasks;
+using OrchardCore.ContentManagement;
+
+namespace DFC.ServiceTaxonomy.GraphSync.Handlers.Interfaces
+{
+    public interface ISyncOrchestrator
+    {
+        Task<bool> SaveDraft(ContentItem contentItem);
+        Task<bool> Publish(ContentItem contentItem);
+        Task<bool> DiscardDraft(ContentItem contentItem);
+    }
+}
